# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THREE NINE POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /ingredients:
    get:
      operationId: getIngredients
      tags: [ingredients]
      summary: List all ingredients
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Ingredients list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ingredient"
    post:
      operationId: createIngredient
      tags: [ingredients]
      summary: Create an ingredient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
  /ingredients/{id}:
    get:
      operationId: getIngredient
      tags: [ingredients]
      summary: Get ingredient by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ingredient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateIngredient
      tags: [ingredients]
      summary: Update an ingredient
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngredientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ingredient"
    delete:
      operationId: deleteIngredient
      tags: [ingredients]
      summary: Delete an ingredient
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        productCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, status, productCount, createdAt, updatedAt]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive]
      required: [name]
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, inactive]
...
    Ingredient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string
        purchasePrice:
          type: number
        currentStock:
          type: number
        minimumStock:
          type: number
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, unit, purchasePrice, currentStock, minimumStock, stockStatus, createdAt, updatedAt]
    IngredientInput:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        purchasePrice:
          type: number
        currentStock:
          type: number
        minimumStock:
          type: number
      required: [name, unit, purchasePrice]
    IngredientUpdate:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        purchasePrice:
          type: number
        currentStock:
          type: number
        minimumStock:
          type: number
...