# lib/api-zod/src/generated/types/categoryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryInputStatus } from './categoryInputStatus';

export interface CategoryInput {
  name: string;
  description?: string;
  status?: CategoryInputStatus;
}