# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryUpdateStatus } from './categoryUpdateStatus';

export interface CategoryUpdate {
  name?: string;
  description?: string;
  status?: CategoryUpdateStatus;
}