# lib/api-zod/src/generated/types/ingredientInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THREE NINE POS API
 * OpenAPI spec version: 0.1.0
 */

export interface IngredientInput {
  name: string;
  unit: string;
  purchasePrice: number;
  currentStock?: number;
  minimumStock?: number;
}